self_train/requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.200
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=9.5.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
PyYAML>=6.0
requests>=2.31.0

# Data processing
pandas>=2.0.0
scikit-learn>=1.3.0
albumentations>=1.3.0

# Visualization
tensorboard>=2.13.0
wandb>=0.15.0

# Evaluation metrics
pycocotools>=2.0.6
scipy>=1.10.0

# Optional: For advanced features
rich>=13.0.0
hydra-core>=1.3.0
omegaconf>=2.3.0 